Renamed `core_function` module to `main_loop`
